# frontend/vue-project/merged-api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://10.239.105.110:11082",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v2/storage/merge": {
      "post": {
        "tags": [
          "upload-v-2-service"
        ],
        "operationId": "completeMultipartUpload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/storage/init": {
      "post": {
        "tags": [
          "upload-v-2-service"
        ],
        "operationId": "initUpload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/upload/merge": {
      "post": {
        "tags": [
          "文件传输（通用服务）"
        ],
        "summary": "合并分片",
        "operationId": "mergeChunks",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/upload/init": {
      "post": {
        "tags": [
          "文件传输（通用服务）"
        ],
        "summary": "初始化上传会话",
        "operationId": "initUpload_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUploadSession"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/storage/listParts": {
      "get": {
        "tags": [
          "upload-v-2-service"
        ],
        "operationId": "listParts",
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uploadId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/storage/Download": {
      "get": {
        "tags": [
          "upload-v-2-service"
        ],
        "operationId": "generateDownloadUrl",
        "parameters": [
          {
            "name": "virtualFileId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ai/stream/chat": {
      "post": {
        "tags": [
          "AI应用"
        ],
        "summary": "AI对话",
        "description": "流式输出并输出引用文档",
        "operationId": "streamChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          }
        }
      }
    },
    "/api/ai/process": {
      "post": {
        "tags": [
          "AI应用"
        ],
        "summary": "处理已有文件",
        "description": "处理当前用户拥有的文档",
        "operationId": "processExistingDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentProcessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MergeRequestV2": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PartInfo"
            }
          }
        }
      },
      "PartInfo": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "integer",
            "format": "int32"
          },
          "etag": {
            "type": "string"
          }
        }
      },
      "InitRequestV2": {
        "required": [
          "fileHash",
          "fileName",
          "fileSize"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "文件名",
            "example": "file.ext"
          },
          "fileSize": {
            "type": "integer",
            "description": "文件大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "fileHash": {
            "type": "string",
            "description": "文件的前1MB的Md5",
            "example": "3b1254bd7f1f14ce0d5caca25626733b"
          },
          "chunkSize": {
            "type": "integer",
            "description": "分块的大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int64"
          },
          "uploadId": {
            "type": "string"
          }
        },
        "description": "初始化上传请求的请求体"
      },
      "InitResponseV2": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string"
          },
          "chunkSize": {
            "type": "integer",
            "format": "int64"
          },
          "presignedUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "uploadedChunks": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "isComplete": {
            "type": "boolean"
          }
        }
      },
      "MergeRequest": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "会话ID",
            "example": "550e8400e29b41d4a716446655440000"
          }
        },
        "description": "合并请求体"
      },
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "通用响应对象"
      },
      "InitRequest": {
        "required": [
          "fileName",
          "fileSize",
          "isFolderUpload",
          "md5",
          "parentId"
        ],
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "文件名",
            "example": "file.ext"
          },
          "relativePath": {
            "type": "string",
            "description": "相对路径",
            "example": "path/to/file.ext"
          },
          "fileSize": {
            "type": "integer",
            "description": "文件大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "md5": {
            "type": "string",
            "description": "文件的前1MB的Md5",
            "example": "3b1254bd7f1f14ce0d5caca25626733b"
          },
          "chunkSize": {
            "type": "integer",
            "description": "分块的大小（字节）",
            "format": "int64",
            "example": 1024
          },
          "parentId": {
            "type": "integer",
            "description": "父目录ID",
            "format": "int64",
            "example": 1001
          },
          "isFolderUpload": {
            "type": "boolean",
            "description": "是否为文件夹上传",
            "example": true
          }
        },
        "description": "初始化上传请求的请求体"
      },
      "ResultUploadSession": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/UploadSession"
          }
        },
        "description": "通用响应对象"
      },
      "UploadSession": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "fileName": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "md5": {
            "type": "string"
          },
          "ossKey": {
            "type": "string"
          },
          "totalChunks": {
            "type": "integer",
            "format": "int32"
          },
          "parentId": {
            "type": "integer",
            "format": "int64"
          },
          "relativePath": {
            "type": "string"
          },
          "isFolderUpload": {
            "type": "boolean"
          },
          "presignedUrl": {
            "type": "string"
          }
        },
        "description": "响应数据"
      },
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应状态码，1表示成功",
            "format": "int32",
            "example": 1
          },
          "msg": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "string",
            "description": "响应数据"
          }
        },
        "description": "通用响应对象"
      },
...
      "ChatRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          }
        }
      },
      "SseEmitter": {
        "type": "object",
        "properties": {
          "timeout": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DocumentProcessRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "用户文件ID",
            "format": "int64",
            "example": 1001
          }
        },
        "description": "RAG处理请求体"
      },
...
    }
  }
}